python_backend/requirements.txt

# Core dependencies for the authentication system
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
requests==2.31.0

# For future Google Sheets integration
google-api-python-client==2.103.0
gspread==5.11.3

# Optional: For more secure password hashing (upgrade from SHA-256)
bcrypt==4.0.1